Combine API server boot op and restart op.
